set up chakra ui

+++ backend/controllers/productController.js
@@ -1,8 +1,8 @@
-import Product from './models/productModel.js';
+import Product from '../models/productModel.js';
 import mongoose from 'mongoose';
 
 
-const getProducts = async (req, res) => {
+export const getProducts = async (req, res) => {
     try {
         const products = await Product.find({});
         res.status(200).json({ success: true, data: products});
@@ -13,7 +13,7 @@ const getProducts = async (req, res) => {
 }
 
 
-const createProduct = async (req, res) => {
+export const createProduct = async (req, res) => {
     const product = req.body;
 
     if(!product.name || !product.price || !product.image){
@@ -32,19 +32,23 @@ const createProduct = async (req, res) => {
 }
 
 
-const deleteProduct = async (req, res) => {
+export const deleteProduct = async (req, res) => {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid()){
+        return res.status(404).json({ success: false, message: "Product not found"});
+    }
 
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     }catch(error){
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
 
 
-const updateProduct = async (req, res) => {
+export const updateProduct = async (req, res) => {
     const { id } = req.params;
 
     const product = req.body;
@@ -59,13 +63,4 @@ const updateProduct = async (req, res) => {
     } catch (error) {
         res.status(500).json({ success: false, message: "Server Error"});
     }
-}
-
-
-
-export default {
-    getProducts,
-    createProduct,
-    deleteProduct,
-    updateProduct,
 }
